dev-merra-data-ingestion/requirements.txt
aniso8601==9.0.1
boto3==1.21.32
botocore==1.24.32
certifi==2021.10.8
cftime==1.6.0
charset-normalizer==2.0.12
click==8.1.2
cloudinary==1.29.0
Flask==2.1.1
Flask-RESTful==0.3.9
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
jmespath==1.0.0
kiwisolver==1.4.2
MarkupSafe==2.1.1
munkres==1.1.4
netCDF4==1.5.8
numpy==1.22.3
pika==1.2.0
Pillow==9.0.1
python-decouple==3.6
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
s3transfer==0.5.2
urllib3==1.26.9
Werkzeug==2.1.0
zipp==3.7.0
matplotlib==3.5.1
Flask==2.1.1
Flask-RESTful==0.3.9
requests==2.27.1
python-decouple==3.6
python-dotenv==0.20.0